pyproject.toml
[tool.poetry]
name = "id-broker"
version = "0.0.1"
description = "authorization SDK used to build protected Web APIs"
authors = ["Yin Long <yangaound@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/yangaound/id-broker"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Operating System :: POSIX",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Intended Audience :: Developers",
]
packages = [
    { include = "id_broker" },
]
[tool.poetry.dependencies]
python = ">=3.8.0,<3.12"
pyjwt = "^2.8.0"
requests-oauthlib = "^1.3.1"
django-cors-headers = "^4.3.1"
django-filter = "^23.4"
djangorestframework = "^3.14.0"
psycopg2-binary = "2.9.5"
gunicorn = "^21.2.0"
django = "4.1.2"

[tool.poetry.group.dev.dependencies]
autoflake = "^1.4"
black = "^22.6.0"
isort = "^5.10.1"
pylint = "^2.14.5"
whitenoise = "^6.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[package.dependencies]
[tool.isort]
multi_line_output = 3
line_length = 120
include_trailing_comma = true

[tool.black]
line_length = 120

[tool.pylint.format]
max-line-length=120
ignore-long-lines = "^\\s*(# )"